Versioning courses works, one bug described below.
For some reason, some courses are having multiple versions inserted in one scrape. This shouldn’t be the case if all course urls are unique and scraped only once.

Potential fix: Ensure all course urls are unique. If this doesn’t work, we can live with this problem for now